Custom exception for application
Throwing various exceptions and more logs is the crux of this commit. More
checks are added.

The ConnectionError used in feeds.py was unknown. To make it known, I had to
import requests.ConnectionError. But feeds.py has no usage of requests library.
So, in order to clean everything up, I made custom exceptions for the
application consisting of
a. ConnectionError - Error connecting to News Server
b. ParseError - Error parsing HTML/XML feed
c. InputError - When argument given to a is not right.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>