Better to have a print_date attribute as well
This would require me to completely remove the old datastore with all its values
and start afresh.

Having a print_date attribute would make it easier to serve it in the client.
Hence, it is made a part of the api as well.

Bumped the version number because of this change in datastore.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>