Half implementation of /api/news
On sending a key to /api/news, the extracted news text must be obtained. For
this, first the link should be obtained. This is done by adding a classmethod to
OpinionList. Also, the kind would also prove useful as different kinds of
opinion (for ex. cartoons, leads etc) have different types of format on the TH
website. Their data have to be extracted in a different manner.

To test the working, currently, a json with key, url and kind values are
returned on receiving a POST request. This would obviously be removed. The
client would not receive any url. The url and kind attributes would then be
passed on to extract script to get the article text.

Test script test/api_news.py currently prints the key, url, kind obtained from
POST request to /api/news.
To get the keys, test/api_list.py has been used.
Needless to say, this was done only for intermediate testing and would not be
present in the final scripts.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>