Implementation for Extract
Extract the html snippet containing the actual article and pass it on as a
response to POST on /api/news.

The Extract class provides a much better rendering than before by using an html
snippet generated by BeautifulSoup and devoid of unnecessary tags, attributes
etc.

Hence, it
Fixes #5
Fixes #8

There are various types of articles. These articles are listed in constants.Kind
Not all of them are supported yet.

Cartoons are irrelevant for my application - even though the code to extract
them has been written.

Letters - might not be useful. get_article works for letters, but it is not needed
now.

Blogs - any request to blogs are being blocked by the server with a 403 response

Readers-Editor - It contains corrections in articles printed before. Irrelevant
for the application

sunday-anchor - any request to sunday-anchor are being blocked by a 404 response

Even though these items are listed, they are not supported.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>