Do not keep Opinions older than a week
The Hindu has policy regarding its data. To play on the safer side, I'll store
the data for at max 7 days after which it will be deleted.
To ensure this, a cron job running clean every 24 hours is made.

The database is queried to find out the day of the latest article. From this,
the midnight of the seventh earlier day is calculated and all entries before
that are deleted.

It is then verified that difference between the first and last article is not
more than 7 days.

But before querying, allow some time so as to finish the deletion process. By
hit and trial, 0.5 second seems to work. On not waiting, datastore replies with
None and hence, DatabaseEmptyError is thrown.

Appropriate constants and errors are added.
Any query to /clean must be admin generated one. Lest anyone deletes the server
data.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>