Kickstart unit testing of extract
Unit testing is an important part of a project.

Beginnig with testing of extract module, we recognise that clean_copy is the
most important function of the module. The functioning of Extract depends solely
on correct clean_copy.

Only the very basic tests are added to start off.

To run the test, one can use pytest or simply python itself.

Using pytest

$ py.test

Simply using python

$ python test_extract1.py

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>