Kinds attribute in request to `/api/list`
If user is interested in only certain kinds of articles, only those should be
sent back.
For this, the post request must contain a 'kinds' attribute containing a list of
kinds they require.

Even though 'kinds' attribute is encouraged, if a user does not send it, the
response contains a list of DEFAULT articles. The DEFAULT type is added to
constants.

Further, 'kinds' can be 'all' or 'default' as well.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>